add dynamic delays

+++ grp_slave/utils/keyboard.py
@@ -2,7 +2,7 @@ import time
 import pyautogui
 
 
-def hold_key(key: str, timeout: int = 5):
+def hold_key(key: str, timeout: float = 5):
     pyautogui.keyDown(key)
     time.sleep(timeout)
     pyautogui.keyUp(key)
